change double to float

+++ open-simplex-noise.h
@@ -40,9 +40,9 @@ struct osn_context;
 int open_simplex_noise(int64_t seed, struct osn_context **ctx);
 void open_simplex_noise_free(struct osn_context *ctx);
 int open_simplex_noise_init_perm(struct osn_context *ctx, int16_t p[], int nelements);
-double open_simplex_noise2(const struct osn_context *ctx, double x, double y);
-double open_simplex_noise3(const struct osn_context *ctx, double x, double y, double z);
-double open_simplex_noise4(const struct osn_context *ctx, double x, double y, double z, double w);
+float open_simplex_noise2(const struct osn_context *ctx, float x, float y);
+float open_simplex_noise3(const struct osn_context *ctx, float x, float y, float z);
+float open_simplex_noise4(const struct osn_context *ctx, float x, float y, float z, float w);
 
 #ifdef __cplusplus
 	}
